persist chats on app level
