+++ src/hooks/use-messages.hook.ts
@@ -1,14 +1,18 @@
 import _ from "lodash";
-import { useState } from "react";
+import { useContext } from "react";
 import { useParams } from "react-router";
+import { AppStateContext } from "../App";
 import { Message } from "../model/message";
 
-export const useMessages = (): [Message[], (newMessage: Message) => void] => {
-  const params = useParams();
-  const chatId = params["chatId"]!;
-  const [messages, setMessages] = useState<Message[]>([]);
-  const addMessage = (newMessage: Message) => {
-    setMessages(currentMessages => [...currentMessages, newMessage]);
-  };
-  return [messages, addMessage];
+export const useMessages = (): [string, Message[], (newMessage: Message) => void] => {
+  const { chats, addMessageToChat } = useContext(AppStateContext);
+  const routeParams = useParams();
+  const chatId = routeParams["chatId"]!;
+  const { name, messages } = _.find(chats, (c) => c.id === chatId)!;
+  const addMessage = _.curry(addMessageToChat)(chatId);
+  return [
+    name,
+    messages,
+    addMessage
+  ];
 };
